+++ src/constants/database.ts
@@ -1,1 +1,3 @@
 export const DEFAULT_DATABASE_FOLDER = "./db";
+export const DEFAULT_MESSAGE_COMMAND_PREFIX = "?";
+export const SLASH_COMMAND_INTENT_PREFIX = "/";