+++ CHANGELOG.md
@@ -5,10 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added a new `/setprefix` command to change the guild's preferred message-command prefix. (This means you can change the default `?` to any 1-3 char string you want!) This feature has been around for a long time, in the form of the now-gone `/config` family of subcommands. Those commands used a storage medium separate from the main SQLite database, and that was always weird to me. Enjoy the new more stable experience!
+
 ### Changed
+- The `/config` command now prints a message instructing the user to use the `/setprefix` command instead.
+- Moved guild-scoped message-command prefix config to the SQLite database. I plan to retain this migration code until the next Semver Major version.
 - Translate duration strings to a language and format that makes sense to the user (for private responses) or to the guild (for public responses).
 - `/quo stats` no longer repsonds publicly, even when the channel of invocation is the queue channel.
 - `/now-playing` has been renamed to `/nowplaying`. The command retains `now-playing` as an alias, in case users wish to do things the hard way.
+
+### Removed
+- Removed legacy `/config` command subtree. Use the `/setprefix` command instead.
 
 ## [1.7.1] - 2022-08-14
 ### Added